# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.ErrorResponse:
    description: Error response returned when a request fails
    properties:
      error:
        example: additional error details
        type: string
      message:
        example: Invalid credentials
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@actalog.com
    name: ActaLog Support
    url: https://github.com/johnzastrow/actalog/issues
  description: ActaLog is a mobile-first CrossFit workout tracker API. It provides
    endpoints for user authentication, workout logging, movement tracking, personal
    records, and administrative functions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://actalog.com/terms/
  title: ActaLog API
  version: 0.17.0
paths:
...
  /performance/movements/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all performance records for a movement with calculated
        1RM values
      parameters:
      - description: Movement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Performance history with best 1RM
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid movement ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get movement performance history
      tags:
      - performance
  /performance/search:
    get:
      consumes:
      - application/json
      description: Search across both movements and WODs in a single query
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Max results per category (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Combined search results
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing search query
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unified search for movements and WODs
      tags:
      - performance
  /performance/wods/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all performance records for a specific WOD
      parameters:
      - description: WOD ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: WOD performance history
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid WOD ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get WOD performance history
      tags:
      - performance
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"